Windows: Refactor sysinfo for compilation
